addInspectorInAppDelegate
