+++ Navigation/Profile/Checker.swift
@@ -5,7 +5,7 @@
 //  Created by kubmakk on 13.01.2025.
 //
 
-import UIKit
+import Foundation
 
 class Checker {
     static let shared = Checker()
@@ -18,4 +18,15 @@ class Checker {
     func check(login: String, password: String) -> Bool {
         login == self.login && password == self.password
     }
+}
+
+protocol LoginViewControllerDelegate{
+    func check(login: String, password: String) -> Bool
+}
+
+struct LoginInspector: LoginViewControllerDelegate{
+    func check(login: String, password: String) -> Bool{
+        return Checker.shared.check(login: login, password: password)
+    }
+
 }
